Separate src/main.js to each panel/functionality with using context
